requirements.txt
Django==5.0.1
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-modeltranslation==0.18.11
django-ckeditor==6.7.0
django-jazzmin==2.6.0